requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.4
Markdown==3.3.7
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0
